+++ src/terraform_manager/entities/workspace.py
@@ -13,7 +13,8 @@ class Workspace:
         auto_apply: bool,
         is_locked: bool,
         working_directory: str,
-        execution_mode: str
+        execution_mode: str,
+        speculative: bool
     ):  # pragma: no cover
         self.workspace_id = workspace_id
         self.name = name
@@ -28,6 +29,7 @@ class Workspace:
         self.is_locked = is_locked
         self.working_directory = working_directory
         self.execution_mode = execution_mode
+        self.speculative = speculative
 
     def is_terraform_version_newer_than(self, version: str) -> bool:
         if self.terraform_version == LATEST_VERSION:
@@ -49,7 +51,7 @@ class Workspace:
             f"Workspace(id={self.workspace_id}, name={self.name}, "
             f"terraform_version={self.terraform_version}, auto_apply={self.auto_apply}, "
             f"is_locked={self.is_locked}, working_directory={self.working_directory}, "
-            f"execution_mode={self.execution_mode})"
+            f"execution_mode={self.execution_mode}, speculative={self.speculative})"
         )
 
     def __str__(self):