+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Added functionality for the `--enable-auto-apply`/`--disable-auto-apply` operations (by [@cooperwalbrun](https://github.com/cooperwalbrun))
+* Added functionality for the `--enable-speculative`/`--disable-speculative` operations (by [@cooperwalbrun](https://github.com/cooperwalbrun))
 
 ### Changed